remove test_mode guard from process_neuron_manual
in order to facilitate the deployment of the icp neurons, the test_mode guard from the process_neuron_manual call is removed. This will allow to manually call the trigger to synchronise the config of the neurons under control of the SNS. Make sure to activate this again once the neurons are setup to avoid any spamming of the canister via this call.